Extract the city giards system from DB.
